Move collision cross section in appropriate file
